added layers, gitignore, readme
