+++ README.md
@@ -1,0 +1,2 @@
+# transformer
+My implementation of the Transformer model (Attention is All You Need, Vaswani et al.)
