+++ layers/attentionBlock.py
@@ -1,16 +1,18 @@
 import torch
-import torch.nn as nn
+from torch import nn
 import math
 from torch import Tensor
 import typing
 from typing import Optional
+from scaledDotProductAttention import ScaledDotProductAttention
 
 
 class MultiHeadAttention(nn.Module):
     def __init__(self, embedding_dimension: int, key_dimension: int, value_dimension: int, num_heads: int):
         super().__init__()
 
-        assert (embedding_dimension % num_heads == 0, "model dimension must be divisible by number of heads.")
+        assert (embedding_dimension % num_heads == 0,
+                "model dimension must be divisible by number of heads.")
         self.model_dimension = embedding_dimension
         self.num_heads = num_heads
         self.key_dimension = key_dimension
@@ -18,10 +20,14 @@ class MultiHeadAttention(nn.Module):
 
         self.scaled_dot_product_attention = ScaledDotProductAttention()
 
-        self.W_q = nn.Linear(embedding_dimension, num_heads * self.key_dimension)  # query weight matrix
-        self.W_k = nn.Linear(embedding_dimension, num_heads * self.key_dimension)  # keys weight matrix
-        self.W_v = nn.Linear(embedding_dimension, num_heads * self.value_dimension)  # values weight matrix
-        self.W_o = nn.Linear(num_heads * self.value_dimension, embedding_dimension)  # output weight matrix
+        self.W_q = nn.Linear(embedding_dimension, num_heads *
+                             self.key_dimension)  # query weight matrix
+        self.W_k = nn.Linear(embedding_dimension, num_heads *
+                             self.key_dimension)  # keys weight matrix
+        self.W_v = nn.Linear(embedding_dimension, num_heads *
+                             self.value_dimension)  # values weight matrix
+        self.W_o = nn.Linear(num_heads * self.value_dimension,
+                             embedding_dimension)  # output weight matrix
 
     def split_heads(self, tensor: Tensor):
         # reshape input to have num_heads for multi-headed attention
@@ -60,32 +66,9 @@ class MultiHeadAttention(nn.Module):
         value = self.split_heads(value)
 
         # get attention matrix
-        attention_matrix = self.scaled_dot_product_attention(query, key, value, mask)
+        attention_matrix = self.scaled_dot_product_attention(
+            query, key, value, mask)
 
         # multiply by output weights to get attention vector
         output = self.W_o(self.combine_heads(attention_matrix))
-        return output
-
-
-class ScaledDotProductAttention(nn.Module):
-    def __init__(self):
-        super().__init__()
-        self.softmax = nn.Softmax(dim=3)
-
-    def forward(self, query: Tensor, key: Tensor, value: Tensor, mask: Optional[int] = None):
-        # input: Tensor[batch_size, head_number, length, key_dimension]
-        # output: Tensor[batch_size, head_number, length, value_dimension]
-
-        batch_size, head_number, length, key_dimension = key.size()
-        # formula to calculate attention score
-        attention_score = query(key.transpose(2, 3) / math.sqrt(key_dimension))
-
-        # handle mask
-        if mask is not None:
-            attention_score = attention_score.masked_fill(mask == 0, -1e9)
-
-        # get weighted sum attention vectors, weighted with respect to attention scores
-        # softmax to normalize, sum = 1
-        attention_probability = self.softmax(attention_score)
-        output = attention_probability(value)
         return output
