+++ layers/positionWiseFeedForward.py
@@ -1,0 +1,8 @@
+from torch import nn
+from torch import Tensor
+
+class PositionWiseFeedForward(nn.Module):
+    def __init__(self, embedding_dimension: int):
+        super().__init__()
+        self.
+
